+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [Opened project's Language Server process is restarted after a crash][14832].
 - [Geo Map Visualization may display GeoSpatial data][14859]
 - [Component height can be resized][14849]
+- [Add right click menu for multiple components][14640].
 - [Execution can be scheduled for the specific version tag][14883]
 
 [14590]: https://github.com/enso-org/enso/pull/14590
@@ -23,6 +24,7 @@
 [14832]: https://github.com/enso-org/enso/pull/14832
 [14859]: https://github.com/enso-org/enso/pull/14859
 [14849]: https://github.com/enso-org/enso/pull/14849
+[14640]: https://github.com/enso-org/enso/pull/14640
 [14883]: https://github.com/enso-org/enso/pull/14883
 
 #### Enso Standard Library